Rewrite the script invocation logic to work around limitations
Boxstarter doesn't work right if the installation of a package then initiates the installation of another package.  In that case, if the child package triggers a reboot, after the reboot the installation of the parent package does not resume.

Thus, I've put back the kludgy "checkpoint" mechanism to detect which scripts have already run in a given install.